# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DWLite Open API",
        "description": "\u7403\u978b\u4f9b\u5e94\u94fe\u8d44\u6e90\u64ae\u5408\u4e0e\u6e20\u9053\u5206\u9500\u5e73\u53f0 Open API\n\n## \u8ba4\u8bc1\u65b9\u5f0f\n\n\u6240\u6709 API \u8bf7\u6c42\u9700\u8981\u4ee5\u4e0b Header:\n- `X-Api-Key`: API Key ID\n- `X-Timestamp`: Unix \u65f6\u95f4\u6233\uff08\u79d2\uff09\n- `X-Nonce`: \u552f\u4e00 UUID\uff08\u9632\u91cd\u653e\uff09\n- `X-Signature`: HMAC-SHA256 \u7b7e\u540d\n\n## \u7b7e\u540d\u7b97\u6cd5\n\n```\ncanonical = METHOD + \"\\n\" + PATH + \"\\n\" + TIMESTAMP + \"\\n\" + NONCE + \"\\n\" + SHA256(BODY)\nsignature = \"sha256=\" + HMAC-SHA256(canonical, API_SECRET)\n```\n\n## \u54cd\u5e94\u683c\u5f0f\n\n\u6210\u529f\u54cd\u5e94:\n```json\n{\n    \"success\": true,\n    \"data\": { ... },\n    \"requestId\": \"uuid\"\n}\n```\n\n\u9519\u8bef\u54cd\u5e94:\n```json\n{\n    \"success\": false,\n    \"error\": {\n        \"code\": \"ERROR_CODE\",\n        \"message\": \"\u9519\u8bef\u63cf\u8ff0\"\n    },\n    \"requestId\": \"uuid\"\n}\n```\n",
        "contact": {
            "name": "DWLite API Support"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1/open",
            "description": "Open API"
        }
    ],
    "paths": {
...
        "/api/v1/open/merchant/inventory": {
            "get": {
                "tags": [
                    "Merchant - Inventory"
                ],
                "summary": "List merchant inventory.",
                "operationId": "get_open_api_merchant_inventory_list",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "sku",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 100,
                            "nullable": true
                        }
                    },
                    {
                        "name": "warehouseCode",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 50,
                            "nullable": true
                        }
                    },
                    {
                        "name": "lowStock",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "nullable": true
                        }
                    },
                    {
                        "name": "outOfStock",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
        "/api/v1/open/warehouse/inventory": {
            "get": {
                "tags": [
                    "Warehouse - Inventory"
                ],
                "summary": "List inventory.",
                "operationId": "get_open_api_warehouse_inventory_list",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "sku",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 100,
                            "nullable": true
                        }
                    },
                    {
                        "name": "merchantCode",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 50,
                            "nullable": true
                        }
                    },
                    {
                        "name": "lowStock",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "nullable": true
                        }
                    },
                    {
                        "name": "outOfStock",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/open/warehouse/inventory/stocktake": {
            "post": {
                "tags": [
                    "Warehouse - Inventory"
                ],
                "summary": "Submit stocktake results.",
                "operationId": "post_open_api_warehouse_inventory_stocktake",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StocktakeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/open/warehouse/inventory/adjustment": {
            "post": {
                "tags": [
                    "Warehouse - Inventory"
                ],
                "summary": "Submit inventory adjustment.",
                "operationId": "post_open_api_warehouse_inventory_adjustment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdjustmentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StocktakeItemDto": {
                "required": [
                    "sku",
                    "countedQuantity"
                ],
                "properties": {
                    "sku": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "countedQuantity": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "shelvingLocation": {
                        "type": "string",
                        "maxLength": 200,
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 500,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "StocktakeRequest": {
                "required": [
                    "items",
                    "countedAt"
                ],
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StocktakeItemDto"
                        },
                        "minItems": 1
                    },
                    "countedAt": {
                        "type": "string"
                    },
                    "counterName": {
                        "type": "string",
                        "maxLength": 100,
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 500,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "AdjustmentItemDto": {
                "required": [
                    "sku",
                    "adjustmentQuantity",
                    "reason"
                ],
                "properties": {
                    "sku": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "adjustmentQuantity": {
                        "type": "integer"
                    },
                    "reason": {
                        "type": "string",
                        "enum": [
                            "damaged",
                            "lost",
                            "found",
                            "correction",
                            "other"
                        ]
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 500,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "AdjustmentRequest": {
                "required": [
                    "items",
                    "adjustedAt"
                ],
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AdjustmentItemDto"
                        },
                        "minItems": 1
                    },
                    "adjustedAt": {
                        "type": "string"
                    },
                    "adjustedBy": {
                        "type": "string",
                        "maxLength": 100,
                        "nullable": true
                    },
                    "notes": {
                        "type": "string",
                        "maxLength": 1000,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key ID",
                "name": "X-Api-Key",
                "in": "header"
            },
            "TimestampAuth": {
                "type": "apiKey",
                "description": "Unix \u65f6\u95f4\u6233\uff08\u79d2\uff09",
                "name": "X-Timestamp",
                "in": "header"
            },
            "NonceAuth": {
                "type": "apiKey",
                "description": "\u552f\u4e00 UUID\uff08\u9632\u91cd\u653e\u653b\u51fb\uff09",
                "name": "X-Nonce",
                "in": "header"
            },
            "SignatureAuth": {
                "type": "apiKey",
                "description": "HMAC-SHA256 \u7b7e\u540d",
                "name": "X-Signature",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "ApiKeyAuth": [],
            "TimestampAuth": [],
            "NonceAuth": [],
            "SignatureAuth": []
        }
    ],
...
}